update: `StoreSettingControllerV2` introduced
