+++ includes/REST/StoreSettingController.php
@@ -2,12 +2,7 @@
 
 namespace WeDevs\Dokan\REST;
 
-use WeDevs\Dokan\Vendor\Settings;
-use WeDevs\Dokan\Vendor\Vendor;
-use WP_HTTP_Response;
 use WP_REST_Controller;
-use WP_REST_Request;
-use WP_REST_Response;
 use WP_REST_Server;
 use WP_Error;
 
@@ -24,7 +19,7 @@ class StoreSettingController extends WP_REST_Controller {
      *
      * @var string
      */
-    protected $namespace = 'dokan/v2';
+    protected $namespace = 'dokan/v1';
 
     /**
      * Route name
@@ -34,20 +29,6 @@ class StoreSettingController extends WP_REST_Controller {
     protected $rest_base = 'settings';
 
     /**
-     * @var Settings
-     */
-    protected $vendor_settings;
-
-    /**
-     * Constructor.
-     *
-     * @return void
-     */
-    public function __construct() {
-        $this->vendor_settings = new Settings();
-    }
-
-    /**
      * Register all routes related to settings
      *
      * @return void
@@ -55,169 +36,58 @@ class StoreSettingController extends WP_REST_Controller {
     public function register_routes() {
         register_rest_route(
             $this->namespace, '/' . $this->rest_base, [
-                [
-                    'methods'             => WP_REST_Server::READABLE,
-                    'callback'            => [ $this, 'get_settings_list' ],
-                    'permission_callback' => [ $this, 'get_settings_permission_callback' ],
-                ],
-            ]
-        );
-        register_rest_route(
-            $this->namespace, '/' . $this->rest_base . '/(?P<group_id>[\w-]+)', [
-                'args'   => [
-                    'group' => [
-                        'description' => __( 'Settings group ID.', 'dokan-lite' ),
-                        'type'        => 'string',
-                    ],
-                ],
 				[
 					'methods'             => WP_REST_Server::READABLE,
-					'callback'            => [ $this, 'get_settings_group' ],
+					'callback'            => [ $this, 'get_settings' ],
 					'permission_callback' => [ $this, 'get_settings_permission_callback' ],
 				],
                 [
                     'methods'             => WP_REST_Server::EDITABLE,
-                    'args'                => $this->update_settings_group_args(),
                     'callback'            => [ $this, 'update_settings' ],
                     'permission_callback' => [ $this, 'get_settings_permission_callback' ],
                 ],
 			]
         );
-        register_rest_route(
-            $this->namespace, '/' . $this->rest_base . '/(?P<group_id>[\w-]+)/(?P<id>[\w-]+)', [
-                'args'   => [
-                    'group' => [
-                        'description' => __( 'Settings group ID.', 'dokan-lite' ),
-                        'type'        => 'string',
-                    ],
-                    'id' => [
-                        'description' => __( 'Settings ID.', 'dokan-lite' ),
-                        'type'        => 'string',
-                    ],
-                ],
-                [
-                    'methods'             => WP_REST_Server::READABLE,
-                    'callback'            => [ $this, 'get_single_settings' ],
-                    'permission_callback' => [ $this, 'get_settings_permission_callback' ],
-                ],
-                [
-                    'methods'             => WP_REST_Server::EDITABLE,
-                    'args'                => $this->update_single_settings_args(),
-                    'callback'            => [ $this, 'update_single_settings' ],
-                    'permission_callback' => [ $this, 'get_settings_permission_callback' ],
-                ],
-            ]
-        );
-        register_rest_route(
-            $this->namespace, '/' . $this->rest_base . '/(?P<group_id>[\w-]+)/(?P<parent_id>[\w-]+)/(?P<id>[\w-]+)', [
-                'args'   => [
-                    'group' => [
-                        'description' => __( 'Settings group ID.', 'dokan-lite' ),
-                        'type'        => 'string',
-                    ],
-                    'parent' => [
-                        'description' => __( 'Settings parent ID.', 'dokan-lite' ),
-                        'type'        => 'string',
-                    ],
-                    'id' => [
-                        'description' => __( 'Settings ID.', 'dokan-lite' ),
-                        'type'        => 'string',
-                    ],
-                ],
-                [
-                    'methods'             => WP_REST_Server::READABLE,
-                    'callback'            => [ $this, 'get_single_settings_field' ],
-                    'permission_callback' => [ $this, 'get_settings_permission_callback' ],
-                ],
-                [
-                    'methods'             => WP_REST_Server::EDITABLE,
-                    'args'                => $this->update_single_settings_args(),
-                    'callback'            => [ $this, 'update_single_settings_field' ],
-                    'permission_callback' => [ $this, 'get_settings_permission_callback' ],
-                ],
-            ]
-        );
     }
 
     /**
      * Update Store
      *
+     * @param \WP_REST_Request $request
+     *
      * @since DOKAN_LITE_SINCE
      *
-     * @param WP_REST_Request $request
-     *
-     * @return WP_Error|WP_REST_Response
+     * @return WP_Error|\WP_REST_Response
      */
     public function update_settings( $request ) {
-        return rest_ensure_response( $this->vendor_settings->save_settings( $request ) );
-    }
+        $vendor   = $this->get_vendor();
+        $params   = $request->get_params();
+        $store_id = dokan()->vendor->update( $vendor->get_id(), $params );
 
-    /**
-     * Update Store
-     *
-     * @since DOKAN_LITE_SINCE
-     *
-     * @param WP_REST_Request $request
-     *
-     * @return WP_Error|WP_REST_Response
-     */
-    public function update_single_settings( $request ) {
-        return rest_ensure_response( $this->vendor_settings->save_single_settings( $request ) );
-    }
+        if ( is_wp_error( $store_id ) ) {
+            return new WP_Error( $store_id->get_error_code(), $store_id->get_error_message() );
+        }
 
-    /**
-     * Update Store single settings field
-     *
-     * @since DOKAN_LITE_SINCE
-     *
-     * @param WP_REST_Request $request
-     *
-     * @return WP_Error|WP_REST_Response
-     */
-    public function update_single_settings_field( $request ) {
-        return rest_ensure_response( $this->vendor_settings->save_single_settings_field( $request ) );
+        $store = dokan()->vendor->get( $store_id );
+
+        do_action( 'dokan_rest_store_settings_after_update', $store, $request );
+
+        $stores_data = $this->prepare_item_for_response( $store, $request );
+        $response    = rest_ensure_response( $stores_data );
+
+        return $response;
     }
 
     /**
      * @param $request
      *
-     * @return WP_Error|WP_HTTP_Response|WP_REST_Response
+     * @return mixed|WP_Error|\WP_HTTP_Response|\WP_REST_Response
      */
-    public function get_settings_list( $request ) {
-        return rest_ensure_response( $this->vendor_settings->list_settings() );
-    }
+    public function get_settings( $request ) {
+        $vendor   = $this->get_vendor();
+        $response = dokan_get_store_info( $vendor->id );
 
-    /**
-     * @param $request
-     *
-     * @return WP_Error|WP_HTTP_Response|WP_REST_Response
-     */
-    public function get_settings_group( $request ) {
-        $group_id = $request->get_param( 'group_id' );
-        return rest_ensure_response( $this->vendor_settings->settings_group( $group_id ) );
-    }
-
-    /**
-     * @param $request
-     *
-     * @return WP_Error|WP_HTTP_Response|WP_REST_Response
-     */
-    public function get_single_settings( $request ) {
-        $group_id = $request->get_param( 'group_id' );
-        $id       = $request->get_param( 'id' );
-        return rest_ensure_response( $this->vendor_settings->single_settings( $group_id, $id ) );
-    }
-
-    /**
-     * @param $request
-     *
-     * @return WP_Error|WP_HTTP_Response|WP_REST_Response
-     */
-    public function get_single_settings_field( $request ) {
-        $group_id = $request->get_param( 'group_id' );
-        $parent_id = $request->get_param( 'parent_id' );
-        $id       = $request->get_param( 'id' );
-        return rest_ensure_response( $this->vendor_settings->single_settings_field( $group_id, $parent_id, $id ) );
+        return rest_ensure_response( $response );
     }
 
     /**
@@ -242,7 +112,7 @@ class StoreSettingController extends WP_REST_Controller {
     /**
      * Get vendor
      *
-     * @return Vendor|WP_Error
+     * @return WP_Error | $vendor
      */
     protected function get_vendor() {
         $current_user = dokan_get_current_user_id();
@@ -261,7 +131,7 @@ class StoreSettingController extends WP_REST_Controller {
      * Prepare links for the request.
      *
      * @param \WC_Data $object Object data.
-     * @param WP_REST_Request $request Request object.
+     * @param \WP_REST_Request $request Request object.
      *
      * @return array Links for the given post.
      */
@@ -282,10 +152,10 @@ class StoreSettingController extends WP_REST_Controller {
      * Prepare a single item output for response
      *
      * @param $store
-     * @param WP_REST_Request $request Request object.
+     * @param \WP_REST_Request $request Request object.
      * @param array $additional_fields (optional)
      *
-     * @return WP_REST_Response $response Response data.
+     * @return \WP_REST_Response $response Response data.
      */
     public function prepare_item_for_response( $store, $request, $additional_fields = [] ) {
         $data     = $store->to_array();
@@ -294,23 +164,5 @@ class StoreSettingController extends WP_REST_Controller {
         $response->add_links( $this->prepare_links( $data, $request ) );
 
         return apply_filters( 'dokan_rest_prepare_store_settings_item_for_response', $response );
-    }
-
-    /**
-     * Args for updating a single setting.
-     *
-     * @return array
-     */
-    private function update_single_settings_args() {
-        return $this->vendor_settings->args_schema_for_save_single_settings();
-    }
-
-    /**
-     * Args for updating a single setting.
-     *
-     * @return array
-     */
-    private function update_settings_group_args() {
-        return $this->vendor_settings->args_schema_for_save_settings_group();
     }
 }
